Add: Mappers for cartridges
I'm not a huge fan of this due to the nested layers of dyns, but I
can deal with that later. One possible end-point might be to
specialize the mappers, and have all of them point into essentially
an in-memory copy of the ROM file. That way we'd drop one layer of
vtables.